build: use vite to deal with frontend HMR and build
