+++ index.html
@@ -24,6 +24,6 @@
       title="Fork me on GitHub"
       >Fork me on GitHub</a
     >
-    <script src="index.js"></script>
+    <script type="module" src="src/app/index.tsx"></script>
   </body>
 </html>
