+++ package.json
@@ -10,13 +10,13 @@
     "coveralls": "jest --coverage && cat ./coverage/lcov.info | coveralls",
     "coverage": "jest --coverage",
     "test": "jest --silent",
-    "copy_artifacts": "mkdir -p dist/app/ && cp src/app/index.html dist/app/index.html && cp artifacts/* dist/app/",
-    "build:app": "yarn copy_artifacts && yarn ts-node esbuild.js app",
-    "build:lambda": "yarn ts-node esbuild.js lambda",
-    "build": "yarn copy_artifacts && yarn ts-node esbuild.js all",
+    "copy_artifacts": "mkdir -p dist/app/ && cp ./index.html dist/app/index.html && cp artifacts/* dist/app/",
+    "build:app": "yarn copy_artifacts && yarn ts-node build.js app",
+    "build:lambda": "yarn ts-node build.js lambda",
+    "build": "yarn copy_artifacts && yarn ts-node build.js all",
     "lint:css": "stylelint './src/app/**/*.tsx'",
     "lint": "prettier --write . '!dist' '!coverage' && yarn eslint src/ && yarn tsc --noemit && yarn lint:css",
-    "serve": "yarn copy_artifacts && yarn ts-node esbuild.js serve",
+    "serve": "yarn copy_artifacts && yarn ts-node build.js serve",
     "deploy": "AWS_DEFAULT_REGION=eu-west-1 yarn build && yarn cdk bootstrap && yarn cdk deploy --require-approval never"
   },
   "lint-staged": {
@@ -156,6 +156,7 @@
     "@types/uuid": "^8.3.4",
     "@typescript-eslint/eslint-plugin": "^5.36.1",
     "@typescript-eslint/parser": "^5.36.1",
+    "@vitejs/plugin-react": "^2.1.0",
     "autoprefixer": "^10.4.8",
     "aws-cdk": "^2.40.0",
     "aws-cdk-lib": "^2.40.0",
@@ -189,6 +190,7 @@
     "ts-jest": "^28.0.8",
     "ts-node": "^10.9.1",
     "typescript": "^4.8.2",
+    "vite": "^3.1.0",
     "yarn-upgrade-all": "^0.7.1"
   }
 }
